# Fix: #1112 URL validation accepts "http:" as valid url